requirements.txt
asgiref==3.8.1
asttokens==2.2.1
backcall==0.2.0
colorama==0.4.6
decorator==5.1.1
Django==5.1.2
django-debug-toolbar==4.2.0
django-extensions==3.2.3
executing==1.2.0
ipython==8.14.0
jedi==0.19.0
matplotlib-inline==0.1.6
parso==0.8.3
pickleshare==0.7.5
Pillow==10.0.0
prompt-toolkit==3.0.39
pure-eval==0.2.2
Pygments==2.16.1
six==1.16.0
sqlparse==0.4.4
stack-data==0.6.2
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
wcwidth==0.2.6
django-filter==21.1
djangorestframework==3.14.0
djoser==2.1.0
python-decouple==3.5
drf-extra-fields==3.2.1
gunicorn==20.1.0
python-dotenv==1.0.1
psycopg2-binary==2.9.10